Split Uplynk integration to internal and external part
